Fixed Xbase partial parsing
Adopted the solution that was used for Xtend, e.g.
minimize the cases where partial parsing is applied
and prefer full reparsed in general
