+++ plugins/org.eclipse.xtext.xbase/.settings/.api_filters
@@ -49,4 +49,12 @@
             </message_arguments>
         </filter>
     </resource>
+    <resource path="src/org/eclipse/xtext/xbase/parser/XbaseNodeModelBuilder.java" type="org.eclipse.xtext.xbase.parser.XbaseNodeModelBuilder">
+        <filter id="571473929">
+            <message_arguments>
+                <message_argument value="NodeModelBuilder"/>
+                <message_argument value="XbaseNodeModelBuilder"/>
+            </message_arguments>
+        </filter>
+    </resource>
 </component>
