+++ plugins/org.eclipse.xtext.xbase/META-INF/MANIFEST.MF
@@ -55,6 +55,7 @@ Export-Package: org.eclipse.xtext.xbase,
  org.eclipse.xtext.xbase.interpreter.impl;x-friends:="org.eclipse.xtext.xbase.tests",
  org.eclipse.xtext.xbase.jvmmodel,
  org.eclipse.xtext.xbase.linking;x-internal:=true,
+ org.eclipse.xtext.xbase.parser;x-friends:="org.eclipse.xtend.core",
  org.eclipse.xtext.xbase.parser.antlr;x-internal:=true,
  org.eclipse.xtext.xbase.parser.antlr.internal;x-internal:=true,
  org.eclipse.xtext.xbase.resource;x-friends:="org.eclipse.xtext.xbase.ui,org.eclipse.xtend.ide,org.eclipse.xtend.ide.common",
